googleCloudFunctions/SaveDataForPlots_requirements.txt
# Function dependencies, for example:
# package>=version
google-cloud-storage>=1.40.0
google-cloud-logging>=2.7.0
pandas>=1.1.5
fsspec==2021.11.1
gcsfs==2021.11.1
branca>=0.5.0